Move `ir.Utils.{relativize,fixFileURI}` to `io.URIUtils`.
There was no need for those functions to be in the `ir` module.
Besides, this removes the dependency of the `test-adapter` on the
`ir` package.
